[Subtitles] Fix subtitles in livestreams with high start timestamps

+++ xbmc/cores/VideoPlayer/Interface/TimingConstants.h
@@ -8,10 +8,15 @@
 
 #pragma once
 
+#include <cstdint>
+
 #define DVD_TIME_BASE 1000000
 #define DVD_NOPTS_VALUE 0xFFF0000000000000
 
-constexpr int DVD_TIME_TO_MSEC(double x) { return static_cast<int>(x * 1000 / DVD_TIME_BASE); }
+constexpr int64_t DVD_TIME_TO_MSEC(double x)
+{
+  return static_cast<int64_t>(x * 1000 / DVD_TIME_BASE);
+}
 constexpr double DVD_SEC_TO_TIME(double x) { return x * DVD_TIME_BASE; }
 constexpr double DVD_MSEC_TO_TIME(double x) { return x * DVD_TIME_BASE / 1000; }
